Change long long to the more standard uint64_t.  This will need to be typedefed for some platforms.
svn path=/trunk/timing/; revision=2190

+++ src/timing/timing.h
@@ -1,7 +1,9 @@
 #ifndef __TIMING_H__
 #define __TIMING_H__
 
-long long timing_get_time(void);
-void timing_sleep(long long sleeptime);
+#include <stdint.h>
+
+uint64_t timing_get_time(void);
+void timing_sleep(uint64_t sleeptime);
 
 #endif  /* __TIMING_H__ */
